move bytes_to_dc to lib
